backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
joblib==1.4.2
scikit-learn==1.5.1
python-dotenv==1.0.1
urllib3==2.2.2
